made necessary changes

+++ ft_strncpy.c
@@ -6,11 +6,13 @@
 /*   By: azhuk <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/07/18 21:50:45 by azhuk             #+#    #+#             */
-/*   Updated: 2016/10/12 10:32:36 by azhuk            ###   ########.fr       */
+/*   Updated: 2016/10/15 17:08:37 by azhuk            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-char	*ft_strncpy(char *dest, char *src, unsigned int n)
+#include "libft.h"
+
+char	*ft_strncpy(char *dest, char *src, size_t n)
 {
 	char	*s2;
